backend/config/database.py
from pymongo import MongoClient
import os
from dotenv import load_dotenv

load_dotenv()

MONGO_URL = os.getenv("MONGO_URL", "mongodb://localhost:27017")

client = MongoClient(MONGO_URL)
db = client["hrms_db"]

# collections
employees_collection = db["employees"]
attendance_collection = db["attendance"]
