backend/requirements.txt
fastapi==0.128.5
uvicorn==0.40.0
pymongo==4.16.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.0
email-validator==2.1.0
dnspython==2.7.0
